% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crt_data}
\alias{crt_data}
\title{Sample Data from the Creativity Task}
\format{
A data frame with 2112 rows and 2 variables:
\describe{
\item{time}{Character String of Seconds since January, 1970}
\item{data}{Participant data logged in the task}
}
}
\source{
Kishida Lab at Wake Forest School of Medicine
}
\usage{
crt_data
}
\description{
A data set with responses generated by a member of the Kishida Lab.
}
\keyword{datasets}
