% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parse.R
\name{get_exit_questions}
\alias{get_exit_questions}
\title{Retrieve Exit Questions from CRT}
\usage{
get_exit_questions(.data)
}
\arguments{
\item{.data}{CRT Data processed by \code{\link{read_crt()}}.}
}
\value{
A data frame with two columns corresponding to the \emph{exit questions}:
\itemize{
\item{time:} The time since January, 1970
\item{data:} Participant data
}
}
\description{
This function returns the data from the exit questions of the CRT.
}
\examples{

get_exit_questions(crt_data)

}
\keyword{internal}
