% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_item_responses.R
\name{get_item_responses}
\alias{get_item_responses}
\title{Get Item Responses for CRT Data}
\usage{
get_item_responses(
  .data,
  items = c("brick", "chair", "cup", "key", "knife", "watch")
)
}
\arguments{
\item{.data}{CRT Data processed by \code{\link{read_crt}}.}

\item{items}{The items to query. Default is \code{c("brick", "chair", "cup", "key", "knife","watch")}.}
}
\value{
It returns a data frame with the following columns: \itemize{
\item{item:} The item a to which the given response corresponds.
\item{response:} The creative use for the item. }
}
\description{
Get Item Responses for CRT Data
}
\examples{

# By default, get all six item responses
get_item_responses(crt_data)

# Only get individual item responses
get_item_responses(crt_data, "cup")
get_item_responses(crt_data, "chair")
get_item_responses(crt_data, "Key")

}
