% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_parse.R
\name{parse_init}
\alias{parse_init}
\title{Initial Parsing of CRT Data}
\usage{
parse_init(.data)
}
\arguments{
\item{.data}{CRT Data processed by \code{\link{read_crt()}}.}
}
\value{
It returns a data frame with the following columns:
\itemize{
\item{time:} The time since January, 1970
\item{prompt:} The category that the individual typed in (e.g. month or object)
\item{prompt_id:} The specific text box the participant filled in
\item{keypress} The key the individual pressed
\item{value} The running value in the text box
}
}
\description{
This function accepts the data output from \code{\link{read_crt()}} and
parses it for the initial values.
}
\examples{

parse_init(crt_data)

}
