% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_read.R
\name{read_crt}
\alias{read_crt}
\title{Read Creativity Task Data}
\usage{
read_crt(path)
}
\arguments{
\item{path}{A character string defining the path to the output of the CRT.}
}
\value{
The raw data downloaded from the task. It is a data frame with two
columns, time and data, and 1875 observations.
}
\description{
Read Creativity Task Data
}
\examples{

#' # Read in the raw data included as part of this package
read_crt(path = system.file("extdata/creativity_raw.txt",
                            package = "creatr")
)

}
